pyproject.toml
[tool.poetry]
name = "AudioClassification"
version = "0.1.0"
description = ""
authors = ["Nikolay Vedernikov <nick.vedernikow@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.10.0"
numpy = "1.21.1"
torchvision = "^0.11.1"
torchaudio = "^0.10.0"
wandb = "^0.12.5"
pandas = "^1.3.4"
PyYAML = "^6.0"
python-dotenv = "^0.19.1"
sklearn = "^0.0"
matplotlib = "^3.4.3"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
